Fix the timeout functionality
It is not good practice to throw exceptions from the signal handler, because signal handler has its own stack.
Instead I used setjmp and longjmp to take the control flow out of the signal handler's stack, so that exception
can be thrown.
It is important for us to be able to set alarm multiple times, because we want to execute multiple queries one by one
and we want to set the timeout value for each of them.